Move tests to right place and fix
