+++ packages/openactive-integration-tests/test/templates/b-req.js
@@ -234,6 +234,7 @@ function createBReqWithBusinessCustomer(data) {
       addressCountry: 'GB',
     },
   };
+  return req;
 }
 
 function createNoAccountId(data) {
